+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Add `badgeColor` prop to `UserMenu` component for
+  `ButtonImageWithTextAndBadge`.
+
 ## [2.46.0] - 2019-12-09
 
 Features: